Moved commenting out of components folder
