+++ client/scss/core.scss
@@ -101,7 +101,7 @@ These are classes for components.
 @import '../src/components/LoadingSpinner/LoadingSpinner';
 @import '../src/components/PublicationStatus/PublicationStatus';
 @import '../src/components/PageExplorer/PageExplorer';
-@import '../src/components/CommentApp/main';
+@import '../src/commenting/main';
 
 // Legacy
 @import 'components/icons';
