+++ client/src/commenting/components/CommentHeader/index.tsx
@@ -1,7 +1,7 @@
 import React, { FunctionComponent, useState, useEffect, useRef } from 'react';
 
-import { gettext } from '../../../../utils/gettext';
-import Icon from '../../../Icon/Icon';
+import { gettext } from '../../../utils/gettext';
+import Icon from '../../../components/Icon/Icon';
 import type { Store } from '../../state';
 
 import { Author } from '../../state/comments';
