+++ client/src/commenting/main.scss
@@ -1,4 +1,4 @@
-@import '../../../scss/settings/variables';
+@import '../../scss/settings/variables';
 
 $color-red-very-dark: #901419;
 $color-amber-0: #e9b04d;
